Create basic classes.
baseRobot is a class that can be used year after year with only
small changes. For now, it has driving by encoders. Eventually,
other methods such as IMU can be used.

CenterStageRobot is this year's competition bot. Because of
the drive train it has, it can strafe, so that was added.

In the modes module, we have our two runOpMode implementations.
The teleop one will have all of the button presses and the
autonomous one will just be a list of instructions and
censor reads.